pyproject.toml
[tool.poetry]
name = "kws-transformer"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.26.2"
torch = {version = "^2.1.1+cu121", source = "pytorch"}
hydra-core = "^1.3.2"
lightning = "^2.1.2"
einops = "^0.7.0"
torchaudio = {version = "^2.1.1+cu121", source = "pytorch"}
sox = "^1.4.1"
nnaudio = "^0.3.2"
matplotlib = "^3.8.2"
wandb = "^0.16.1"
scikit-learn = "^1.3.2"
pyaudio = "^0.2.14"
onnx = "^1.15.0"
onnxruntime = "^1.16.3"
dvc = "^3.48.1"
dvc-gdrive = "^3.0.1"


[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
